eip712 first pass
