+++ pallets/quantum-portal/Cargo.toml
@@ -24,6 +24,7 @@ parking_lot = "0.11"
 scale-info = { default-features = false, features = ['derive'], version = '2.1.2' }
 serde = { version = '1.0.130', default-features = false, features = ['derive'] }
 serde_json = { version = '1.0.67', default-features = false, features = ['alloc'] }
+eip712 = "0.1.0"
 
 # Substrate packages
 byte-slice-cast = { version = "1.2.1", default-features = false, features = [] }
